Resolved some build issues on Win32
git-svn-id: https://shawn.svn.sourceforge.net/svnroot/shawn@14 3422d224-1e22-0410-b8a2-c109bc27950b

+++ src/apps/topology/point_gen/lattice_point_gen.cpp
@@ -40,7 +40,7 @@ namespace topology
       throw( std::runtime_error )
    {
       PointGenerator::pre_generate(sc,brh);
-	  spacing_ = sc.environment().optional_double_param("spacing", 1.0);
+	  spacing_ = (float) sc.environment().optional_double_param("spacing", 1.0);
 
       bbox_=extract_box(brh);
       check_fail_count_=true;
